refactor: move colours function and readd to hero
