+++ src/components/Hero/PageHero/Hero.stories.tsx
@@ -1,38 +1,7 @@
 import { Meta, StoryObj } from '@storybook/react'
 import Hero from './Hero'
 import React from 'react'
-import colors from '../../../theme/foundations/colors'
-
-const newColors = colors.brand
-
-// Creates an array of the colour codes, setting colours with objects as values to be a string for the colour code of 500
-
-const colorValuesArray = () => {
-  Object.keys(newColors).forEach((key: string) => {
-    const currentValue = newColors[key as keyof typeof newColors]
-    if (typeof currentValue !== 'string') {
-      newColors[key as keyof typeof newColors] = currentValue['500']
-    }
-  })
-
-  return Object.values(newColors)
-}
-
-// Creates the object where they keys and values are reversed to generate labels. See these docs https://storybook.js.org/docs/react/essentials/controls
-
-const newColourObject = Object.fromEntries(
-  Object.entries(newColors).map(([key, value]) => [value, key])
-)
-
-// generates the argType for textColour and backgroundColour
-
-const colourArgType = {
-  options: colorValuesArray(),
-  control: {
-    type: 'select',
-    labels: newColourObject,
-  },
-}
+import { selectColours } from '../../../atoms/Colours/selectColours'
 
 const meta: Meta = {
   component: Hero,
@@ -53,8 +22,8 @@ export const Default: Story = {
       'Lorem ipsum dolor sit amet consectetur. Sed hac laoreet risus iaculis donec. In dolor tincidunt sed elementum cursus rhoncus adipiscing. Semper sed eget dignissim.',
   },
   argTypes: {
-    backgroundColour: colourArgType,
-    textColour: colourArgType,
+    backgroundColour: selectColours,
+    textColour: selectColours,
   },
   render: (args) => <Hero {...args} />,
 }
